Add 'headers' to 'Todos'/'Projects'
Add header divs to both Todos and Projects at the top of each section.
Create 'Add'  buttons next to each section header (Todos, Projects) to
give ability to 'call' a modal. Both Todos and Projects will call a
modal with a form in order to generate either a new Todo or a new Project