Construct renderTodosToDisplay function
Add renderTodosToDisplay(). Whenever a project button is clicked in
the project's data-section, it will call renderTodosToDisplay. This
function will loop through the project's todoArr and create a div
for each todo which will appear in the todo-display section.

Each div has a checkbox (for toggling whether the todo is active or complete),
and also prints the 'title' of the todo and it's chosen date. The div
also contains 'Edit' and 'Delete' buttons for editing the todo or
deleting it altogether.