Enable todo list to refresh upon adding new todo
Add getProjObj() and updateTodoList() to todo's submit button to enable
automatic refreshes of the todo display. Before this, new todos could
only be seen if user re-clicked the project button after adding the new
todo. Now, the project's todos are automatically shown once a new todo
is added with the 'submit' button