Refactor new edit-todo modal - except for save btn
Refactor majority of the edit todo modal by querying the DOM for the
static modal, injecting the current todo values into the modal, attaching
input event listeners to the form fields, and dynamically adding projects
to the project dropdown.

This is a replacement of the former edit-todo modals which were dynamically
generated each time a new todo was created. This caused a lot of unnecessary
complexity. With this replacement, the todos all share the same modal and
populate the modal with their specific datas as they are called