Prep for refactor of editTodo modal's edit button
The 'edit' button in the editTodo modal is not following
the single responsbility principle and causing a bug to form within the
save button. This bug duplicates todos, etc. This commit is to save the
current codebase before refactor of the 'edit' button's event listener.

Next steps are to parse this event listener and remove unnecessary
responsibility from the editTodo modal's 'edit' button