Refactor function to fix todo duplication bug
Refactor addFunctionalityToTodoEditSaveBtn(). The way this was originally
constructed used a reference of a todo object and attempted to update the
reference before using it to change the real todo within projArr. Now, the
projArr is directly accessed, the appropriate todo found (using the todo's id),
and the todo object is updated directly