Prepare for project ID refactor
Add comment to punchlist at bottom of page in preparation for refactor
of todo creation. Chosen project ID needs to be added to the todo object
upon creation. Currently moving data based on the todo.project matching
the project.name. This is bad practice. Need to move data based on the
UUID.

Merging this snapshot to main branch. That way, if refactor goes awry, will
have a clean slate to redo.