Refactor todo edit btn listener / todo ID storage
Refactor todo edit button listener to pass in both the todo's id and the
project id currently associated with the todo. Before refactor, only the
todo's UUID was stored.

Refactor the ID storage function to 'storeTodoIds' and create separate
setter and getter for the project ID that's associated with the todo