Refactor submitTodo and getProjObj to use projId
Refactor helper functions within the submitTodoButton event listener.
submitTodo was returning the project name, and getProjObj was using the
name to search for the correct project.

Now, submitTodo returns the new todo with the projId as a property. This
allows getProjObj to search for the project based on its UUID rather than
its name. This will scale better and reduce the chance of bugs