Refactor updateTodoList to check ids before render
Todos were rendering the to the todo display whether or not their project
was currently on display. From a UX perspective, this would cause confusion
with the user because, even though they placed a todo in 'Home', for example,
it would still render in 'Finance' if 'Finance' was the project currently
on display.

Use ids to check the current project on display before rendering the todos