Add functions to clear project with del button
Add removeProjFromProjSection() and checkAndClearProjFromDisplay().
removeProjFromProjSection deletes the targeted project from the 'projects'
section in the lefthand sidebar. checkAndClearProjFromDisplay() checks
to see if the project that is to be deleted is currently on display.
If it is - project's todos and header are both deleted from the display.
If another project is currently on display, no changes will take place
within the main display