Refactor submitTodoButton listener - add filters
Refactor the submitTodoButton listener to take filters into account.
Previously, if a filter header & filtered todolist were showing in the UI,
the submitTodoButton would not take that into account. This made it so
the todo would not add to the filtered list (No Due Date, etc.) even if
that todo met the reqs of the filter. The user would have to click off
the list and click back to filter it.