Include dataAttr check in proj and todo del btns
Include a display check for a 'filter' data attribute in both the todo's/
project's delete button event listeners. If a filter such as "All Todos"
was showing on the display, todos would not delete in real time - the user
would have to click away to another proj or filter and click back to
original filter to reflect changes. Same with Projects. If a project was
deleted, its todos would still be showing if a filter was being displayed.
This has now been debugged