Add addTimestampToDate() function
Add function to tack a timestamp onto the date from the new todo date
picker. Though the user's chosen date was appearing in string form, the
lack of a timestamp was causing issues when rendering the todo to the UI.
The UTC time difference (UTC is seven hours ahead of my local time) was
causing the todo's date to render as one day behind the actual chosen date.