Debug timestamp issue within date
Refactor currentDateOrTime(), createDueDateElement(), and remove
addTimestampToDate().

There was still a date discrepancy after the last commit as I was
attempting to add a current timestamp to a new todo, then match it
perfectly with the current date's timestamp. Now, these functions are
constructed in such a way that they 'borrow' either the date or timestamp
of the current date based on what's needed at the time. For exampl, the
timestamp is needed when creating the new todo so that the todo can
render to the DOM correctly