Give fn ability to remove proj/todo from storage
Add functionality to removeFromLocalStorage. This fn will discern whether
the object is an entire project or a single todo. If the object is an
entire project, both the project and all of its todos will be removed
from localStorage. If the obj is a single todo, only that todo will be
removed from localStorage.