Refactor fns to create proj/todo arrs in storage
Refactor "saveToLocalStorage()". Previously, this would add a project or
todo to localStorage in no particular order. Now, it routes the object
based on whether it's a todo or a project, then checks for a todo/proj
arr within localStorage to place the object. If it can't find an array,
it makes an array, then stores the object. This keeps projs and todos
organized upon reloading the page

Refactor getLocalStorageObjs() and editTodoInLocalStorage() to reflect
above changes.